added changes

+++ resources/views/categories.blade.php
@@ -736,8 +736,7 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
-
+                        var quantity = parseInt(item.quantity);
                         var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
@@ -773,8 +772,7 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
-
+                        var quantity = parseInt(item.quantity);
                         var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
@@ -881,9 +879,8 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
-
-                        var subtotal = parseInt(price * quantity);
+                        var quantity = parseInt(item.quantity);
+                        var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
                         $('#subtotal_' + product_id).html('$' + subtotal);
@@ -916,9 +913,8 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
-
-                        var subtotal = parseInt(price * quantity);
+                        var quantity = parseInt(item.quantity);
+                        var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
                         $('#subtotal_' + product_id).html('$' + subtotal);
