+++ resources/views/partials/similar_products_slider.blade.php
@@ -508,7 +508,7 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
+                        var quantity = parseInt(item.quantity);
 
                         var subtotal = parseInt(price * quantity);
                         var cart_total = cart_total + subtotal;
@@ -543,9 +543,9 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
+                        var quantity = parseInt(item.quantity);
 
-                        var subtotal = parseInt(price * quantity);
+                        var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
                         $('#subtotal_' + product_id).html('$' + subtotal);
